Skip completed runs
- Additionally added an explicit rerun option to force a run.

+++ run.sh
@@ -30,6 +30,7 @@ function help() {
     printf "  -%-1s, --%-10s %-60s\n" "n" "neko" "Look for examples in neko."
     printf "  -%-1s, --%-10s %-60s\n" "s" "submit" "Submit the examples to a cluster."
     printf "  -%-1s, --%-10s %-60s\n" " " "dry-run" "Dry run the script."
+    printf "  -%-1s, --%-10s %-60s\n" "r" "re-run" "Re-run the examples."
 
     printf "\n\e[4mAvailable case files:\e[0m\n"
     for case in $(find $EPATH -name "*.case" 2>/dev/null); do
@@ -48,10 +49,11 @@ NEKO=false
 DELETE=false
 CLUSTER=""
 DRY=false
+RERUN=false
 
 # List possible options
-OPTIONS=all,clean,help,neko,delete,submit:,dry-run
-OPT="a,c,h,n,s:,d"
+OPTIONS=all,clean,help,neko,delete,submit:,dry-run,re-run
+OPT="a,c,h,n,s:,d,r"
 
 # Parse the inputs for options
 PARSED=$(getopt --options=$OPT --longoptions=$OPTIONS --name "$0" -- "$@")
@@ -67,6 +69,7 @@ while true; do
     "-d" | "--delete") DELETE=true && shift ;;    # Delete previous runs
     "-s" | "--submit") CLUSTER="$2" && shift 2 ;; # Submit to the queue
     "--dry-run") DRY=true && shift ;;             # Dry run
+    "-r" | "--re-run") RERUN=true && shift ;;     # Re-run the examples
 
     # End of options
     "--") shift && break ;;
@@ -277,6 +280,9 @@ function Submit() {
     cd $LPATH/$example
     if [ $CLUSTER == "DTU" ]; then
         export BSUB_QUIET=Y
+        if [ ! -z "$(bjobs -J $1 2>/dev/null)" ]; then
+            bkill -J $1 1>/dev/null 2>/dev/null
+        fi
         bsub -J $1 -env "all" <job_script.sh
 
     elif [ $CLUSTER == "MN5" ]; then
@@ -327,6 +333,11 @@ for case in ${example_list[@]}; do
         example=$example/$case_name
     fi
 
+    if [ "$RERUN" == false ] && [ -d "$RPATH/$example" ]; then
+        printf '\t%-12s %-s\n' "Skipped:" "$example"
+        continue
+    fi
+
     export log=$LPATH/$example && mkdir -p $log
     [ "$CLEAN" = true ] && rm -fr $log/*
 
@@ -337,6 +348,9 @@ for case in ${example_list[@]}; do
 
         [ ! -z "$CLUSTER" ] && printf '\t%-12s %-s\n' "Queued:" "$example"
         QUEUE="$QUEUE $example"
+        continue
+    elif [ -f "$log/output.log" ]; then
+        printf '\t%-12s %-s\n' "Skipping:" "$example"
         continue
     fi
